Add FlaskConfig field
On dump, the field sources its data from the configured Flask config value.

By default, the field is dump-only (read-only), but it can be configured to update the configuration on load. This is however not smart enough to deal with nested structures and a limited view on the existing configuration value.