Apply `filterwarnings = error` to tests and update
This applies the error filter to force test failures on
warnings, and does the necessary cleanup. Primarily, that means
identifying the two classes of `ignore:` rules which are necessary to
handle cases outside of this library's control. Secondarily, it forces
the deprecation of the older url_for field specification syntax, which
cannot be supported cleanly. Since this was deprecated in the last
release, it is presumably safe to remove it, but the change is still
noted as breaking in the changelog.