Bump marshmallow-sqlalchemy and suppress a warning
This is a move towards better compatibility with sqlalchemy 1.4 / 2.0

Because marshmallow-sqlalchemy has several changes and fixes over its
version range to be compatible with the newer sqlalchemy versions,
push towards the latest version.

Ignore the one warning in the build which we can't get away from right
now. (Requires marshmallow-sqlalchemy v0.29.0, not yet released.)